calculate delta for relations in every relational field-type to prevent deadlock

+++ models/DataObject/Localizedfield/Dao.php
@@ -556,13 +556,20 @@ class Dao extends Model\Dao\AbstractDao
                     '/'.$context['containerType'].'~'.$containerName.'/'.$index.'/%'
                 ).$dirtyLanguageCondition;
 
-            $this->db->executeStatement('DELETE FROM object_relations_'.$object->getClassId() . ' WHERE ' . $sql);
+            if ($deleteQuery) {
+                $this->db->executeStatement('DELETE FROM object_relations_'.$object->getClassId().' WHERE '.$sql);
+            }
 
             return true;
         }
 
-        $sql = 'ownertype = "localizedfield" AND ownername = "localizedfield" and src_id = '.$this->model->getObject()->getId().$dirtyLanguageCondition;
-        $this->db->executeStatement('DELETE FROM object_relations_'.$this->model->getObject()->getClassId() . ' WHERE ' . $sql);
+        if ($deleteQuery) {
+            $sql = 'ownertype = "localizedfield" AND ownername = "localizedfield" and src_id = '.$this->model->getObject(
+                )->getId().$dirtyLanguageCondition;
+            $this->db->executeStatement(
+                'DELETE FROM object_relations_'.$this->model->getObject()->getClassId().' WHERE '.$sql
+            );
+        }
 
         return false;
     }
